Extract ParseError to separate module
